requirements.txt
PyQt6~=6.7.1
ldap3~=2.9.1
python-dotenv~=1.0.1
pip~=23.2.1
